Prevent double encoding of oauth_token in get_authorize_url
urllib.urlencode automatically applies urllib.quote_plus to all keys
and values, so applying urllib.quote before hand will result in double
encoding of the token.

This double encoding doesn't cause a problem, of course, if the token
consists of regular characters. But an equals sign (=), for example,
would get encoded once to "%3D", and then again to "%253D".